# docs/swagger.yaml
basePath: /
definitions:
  admin.AdjustmentRequest:
    properties:
      amount:
        type: string
      direction:
        type: string
      reason:
        type: string
      reference:
        type: string
    type: object
  admin.AdminUserRequest:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  admin.AnalyticsResponse:
    properties:
      active_players:
        type: integer
      bet_count:
        type: integer
      daily:
        items:
          $ref: '#/definitions/admin.DailyMetrics'
        type: array
      deposits:
        example: "120000.00"
        type: string
      ggr:
        example: "6000.00"
        type: string
      house_margin_percent:
        example: "6.00"
        type: string
      rtp_percent:
        example: "94.00"
        type: string
      total_payouts:
        example: "94000.00"
        type: string
      total_wagered:
        example: "100000.00"
        type: string
      withdrawals:
        example: "30000.00"
        type: string
    type: object
  admin.DailyMetrics:
    properties:
      active_players:
        type: integer
      bet_count:
        type: integer
      day:
        example: "2026-09-26"
        type: string
      deposits:
        type: string
      ggr:
        example: "6000.00"
        type: string
      payouts:
        example: "94000.00"
        type: string
      rtp_percent:
        example: "94.00"
        type: string
      wagered:
        example: "100000.00"
        type: string
      withdrawals:
        type: string
    type: object
  admin.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  admin.OverviewResponse:
    properties:
      active_bets:
        type: integer
      active_players:
        type: integer
      active_users:
        type: integer
      bet_count:
        type: integer
      deposits:
        type: string
      ggr:
        type: string
      house_margin_percent:
        type: string
      pending_deposits:
        type: integer
      pending_withdrawals:
        type: integer
      rtp_percent:
        type: string
      total_payouts:
        type: string
      total_player_balances:
        example: "450000.00"
        type: string
      total_users:
        type: integer
      total_wagered:
        type: string
      withdrawals:
        type: string
    type: object
...
  admin.RoleRequest:
    properties:
      role:
        type: string
    type: object
  admin.StatusRequest:
    properties:
      reason:
        type: string
      status:
        type: string
    type: object
...
  game.GameRound:
    properties:
      betting_closes_at:
        type: string
      betting_opened_at:
        type: string
      client_seed:
        type: string
      crash_point:
        type: number
      created_at:
        type: string
      ended_at:
        type: string
      house_edge:
        type: number
      id:
        type: integer
      nonce:
        type: integer
      round_number:
        type: integer
      server_seed:
        type: string
      server_seed_hash:
        type: string
      started_at:
        type: string
      status:
        $ref: '#/definitions/game.RoundStatus'
    type: object
  game.RoundStatus:
    enum:
    - CREATED
    - BETTING_OPEN
    - BETTING_CLOSED
    - RUNNING
    - CRASHED
    - SETTLED
    type: string
    x-enum-varnames:
    - RoundCreated
    - RoundBettingOpen
    - RoundBettingClosed
    - RoundRunning
    - RoundCrashed
    - RoundSettled
  game.Snapshot:
    properties:
      current_multiplier:
        type: string
      phase:
        type: string
      running:
        $ref: '#/definitions/game.GameRound'
      seconds_remaining:
        type: integer
      server_time:
        type: string
      upcoming:
        $ref: '#/definitions/game.GameRound'
    type: object
...
host: localhost:7000
info:
  contact: {}
  description: HTTP API for the Aviator real-money game backend.
  title: Aviator Backend API
  version: "1.0"
paths:
  /api/admin/analytics:
    get:
      description: Requires an active ADMIN role. Analytics use settled rounds, exclude
        cancelled bets, and bucket days in UTC.
      parameters:
      - description: Inclusive RFC3339 timestamp
        in: query
        name: from
        type: string
      - description: Exclusive RFC3339 timestamp
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.AnalyticsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Admin Analytics
      tags:
      - admin
...
  /api/admin/config:
    get:
      description: Requires an active ADMIN role. Analytics use settled rounds, exclude
        cancelled bets, and bucket days in UTC.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Admin Config
      tags:
      - admin
...
  /api/admin/game/current:
    get:
      description: Requires ADMIN; uses the public redaction rules for fairness data.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/game.Snapshot'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get running and upcoming rounds for admin monitoring
      tags:
      - admin
  /api/admin/game/status:
    get:
      description: Requires an active ADMIN role. Analytics use settled rounds, exclude
        cancelled bets, and bucket days in UTC.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Admin System
      tags:
      - admin
  /api/admin/overview:
    get:
      description: Requires an active ADMIN role. Analytics use settled rounds, exclude
        cancelled bets, and bucket days in UTC.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.OverviewResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Admin Overview
      tags:
      - admin
...
  /api/admin/rounds/{id}:
    get:
      description: Requires an active ADMIN role. Analytics use settled rounds, exclude
        cancelled bets, and bucket days in UTC.
      parameters:
      - description: Resource ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Admin Round
      tags:
      - admin
  /api/admin/session:
    get:
      description: Requires an active ADMIN role. Analytics use settled rounds, exclude
        cancelled bets, and bucket days in UTC.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Admin Session
      tags:
      - admin
...
  /api/admin/users/{id}:
    get:
      description: Requires an active ADMIN role. Analytics use settled rounds, exclude
        cancelled bets, and bucket days in UTC.
      parameters:
      - description: Resource ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Admin User
      tags:
      - admin
  /api/admin/users/{id}/role:
    patch:
      consumes:
      - application/json
      description: Requires ADMIN. Administrators can change another user's role but
        cannot change their own role or demote the last active administrator.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: New role
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/admin.RoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Change a user's role
      tags:
      - admin
  /api/admin/users/{id}/status:
    patch:
      consumes:
      - application/json
      description: Requires an active ADMIN role. Analytics use settled rounds, exclude
        cancelled bets, and bucket days in UTC.
      parameters:
      - description: Resource ID
        in: path
        name: id
        required: true
        type: integer
      - description: Action details. Reuse the wallet reference on retries.
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/admin.StatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Admin Status
      tags:
      - admin
  /api/admin/users/{id}/wallet-adjustments:
    post:
      consumes:
      - application/json
      description: Requires an active ADMIN role. Analytics use settled rounds, exclude
        cancelled bets, and bucket days in UTC.
      parameters:
      - description: Resource ID
        in: path
        name: id
        required: true
        type: integer
      - description: Action details. Reuse the wallet reference on retries.
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/admin.AdjustmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Admin Adjust
      tags:
      - admin
  /api/admin/users/admins:
    post:
      consumes:
      - application/json
      description: Requires ADMIN. Creates a new active ADMIN account and records
        the action in the audit log.
      parameters:
      - description: Administrator account details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/admin.AdminUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create an administrator account
      tags:
      - admin
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer" followed by a space and your JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"